Legacy Sapsar Preview (Sapsar 0.0.0)
When Sapsar was only starting out, it was originally called FalconJS, after the fastest bird on the planet. But then, the npm package name was taken, the domains were taken and it sounded too much like I was American.

Sapsar is named after an inner national train called "Sapsan" which is one of the fastest operating trains in Russia.